refactor(voice): extract audio frame codec
